refactor: SignatureMatrixのビルダの実装を分離する
